# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CoopDigital — API de gestion des coopératives cacaoyères
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /api/campagnes/active:
    get:
      operationId: getCampagneActive
      tags: [campagnes]
      summary: Campagne en cours (statut ouverte)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Campagne active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campagne"
        "404":
          description: Aucune campagne active
  /api/campagnes:
    get:
      operationId: listCampagnes
      tags: [campagnes]
      summary: Liste de toutes les campagnes
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Liste campagnes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campagne"
    post:
      operationId: createCampagne
      tags: [campagnes]
      summary: Créer une nouvelle campagne
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampagneInput"
      responses:
        "201":
          description: Campagne créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campagne"
  /api/campagnes/{id}/fermer:
    put:
      operationId: fermerCampagne
      tags: [campagnes]
      summary: Clôturer une campagne
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FermerCampagneInput"
      responses:
        "200":
          description: Campagne fermée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campagne"
...
  /api/fournisseurs/search:
    get:
      operationId: searchFournisseurs
      tags: [fournisseurs]
      summary: Recherche autocomplete fournisseurs
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Résultats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fournisseur"
  /api/fournisseurs/rapport-type:
    get:
      operationId: getRapportTypeFournisseur
      tags: [fournisseurs]
      summary: Répartition des fournisseurs par type
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Rapport
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [typeFournisseur, count]
                  properties:
                    typeFournisseur:
                      type: string
                    count:
                      type: integer
  /api/fournisseurs:
    get:
      operationId: listFournisseurs
      tags: [fournisseurs]
      summary: Liste des fournisseurs
      security: [{ bearerAuth: [] }]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [membre, pisteur, externe]
        - name: section
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Liste fournisseurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fournisseur"
    post:
      operationId: createFournisseur
      tags: [fournisseurs]
      summary: Créer un fournisseur (pisteur/externe)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FournisseurInput"
      responses:
        "201":
          description: Fournisseur créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fournisseur"
  /api/fournisseurs/{id}:
    get:
      operationId: getFournisseurById
      tags: [fournisseurs]
      summary: Fiche fournisseur
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fournisseur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fournisseur"
    put:
      operationId: updateFournisseur
      tags: [fournisseurs]
      summary: Modifier un fournisseur
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FournisseurInput"
      responses:
        "200":
          description: Fournisseur mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fournisseur"
  /api/fournisseurs/depuis-membre/{id}:
    post:
      operationId: createFournisseurDepuisMembre
      tags: [fournisseurs]
      summary: Créer un fournisseur depuis un membre existant
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Fournisseur créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fournisseur"
...
components:
  schemas:
...
    Campagne:
      type: object
      required: [id, cooperativeId, libelle, anneeDebut, anneeFin, dateOuverture, statut, createdAt]
      properties:
        id:
          type: integer
        cooperativeId:
          type: integer
        libelle:
          type: string
        anneeDebut:
          type: integer
        anneeFin:
          type: integer
        dateOuverture:
          type: string
          format: date
        dateFermeture:
          type: ["string", "null"]
          format: date
        statut:
          type: string
          enum: [ouverte, fermee]
        createdAt:
          type: string
          format: date-time
    FermerCampagneInput:
      type: object
      properties:
        dateFermeture:
          type: string
          format: date
    CampagneInput:
      type: object
      required: [libelle, anneeDebut, anneeFin, dateOuverture]
      properties:
        libelle:
          type: string
        anneeDebut:
          type: integer
        anneeFin:
          type: integer
        dateOuverture:
          type: string
          format: date
...
    Fournisseur:
      type: object
      required: [id, cooperativeId, typeFournisseur, code, nom, actif, createdAt]
      properties:
        id:
          type: integer
        cooperativeId:
          type: integer
        typeFournisseur:
          type: string
          enum: [membre, pisteur, externe]
        membreId:
          type: ["integer", "null"]
        code:
          type: string
        nom:
          type: string
        prenoms:
          type: ["string", "null"]
        sexe:
          type: ["string", "null"]
        telephone:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
        nationalite:
          type: ["string", "null"]
        numeroCni:
          type: ["string", "null"]
        origine:
          type: ["string", "null"]
        dateAdhesion:
          type: ["string", "null"]
          format: date
        lieuNaissance:
          type: ["string", "null"]
        actif:
          type: boolean
        createdAt:
          type: string
          format: date-time
    FournisseurInput:
      type: object
      required: [typeFournisseur, nom]
      properties:
        typeFournisseur:
          type: string
          enum: [pisteur, externe]
        nom:
          type: string
        prenoms:
          type: string
        sexe:
          type: string
        telephone:
          type: string
        section:
          type: string
        nationalite:
          type: string
        numeroCni:
          type: string
        origine:
          type: string
        dateAdhesion:
          type: string
          format: date
        lieuNaissance:
          type: string