# lib/api-zod/src/generated/types/campagneInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoopDigital — API de gestion des coopératives cacaoyères
 * OpenAPI spec version: 0.1.0
 */

export interface CampagneInput {
  libelle: string;
  anneeDebut: number;
  anneeFin: number;
  dateOuverture: Date;
}